# backend/src/main/java/uz/developer/zohidjon/backend/auth/AuthenticationResponse.java
package uz.developer.zohidjon.backend.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}