# backend/src/main/java/uz/developer/zohidjon/backend/book/BorrowedBookResponse.java
package uz.developer.zohidjon.backend.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {

    private Integer id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}