# backend/src/main/java/uz/developer/zohidjon/backend/feedback/FeedbackResponse.java
package uz.developer.zohidjon.backend.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}